+++ client/web/src/enterprise/insights/core/backend/code-insights-backend.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs'
 
-import { Insight, InsightDashboard, InsightsDashboardOwner } from '../types'
+import { Insight, InsightsDashboardOwner } from '../types'
 
 import {
     AssignInsightsToDashboardInput,
@@ -12,13 +12,11 @@ import {
     FindInsightByNameInput,
     GetBuiltInsightInput,
     GetLangStatsInsightContentInput,
-    GetSearchInsightContentInput,
     InsightCreateInput,
     InsightUpdateInput,
     RemoveInsightFromDashboardInput,
     CategoricalChartContent,
     SeriesChartContent,
-    UiFeaturesConfig,
     InsightContent,
     InsightPreviewSettings,
     BackendInsightDatum,
@@ -31,20 +29,10 @@ import {
  */
 export interface CodeInsightsBackend {
     /**
-     * Returns all accessible code insights dashboards for the current user.
-     * This includes virtual (like "all insights") and real dashboards.
-     */
-    getDashboards: () => Observable<InsightDashboard[]>
-
-    getDashboardById: (input: { dashboardId: string | undefined }) => Observable<InsightDashboard | null>
-
-    /**
      * Returns all possible visibility options for dashboard. Dashboard can be stored
      * as private (user subject), org level (organization subject) or global (site subject)
      */
     getDashboardOwners: () => Observable<InsightsDashboardOwner[]>
-
-    findDashboardByName: (name: string) => Observable<InsightDashboard | null>
 
     createDashboard: (input: DashboardCreateInput) => Observable<DashboardCreateResult>
 
@@ -91,11 +79,6 @@ export interface CodeInsightsBackend {
     getBuiltInInsightData: (input: GetBuiltInsightInput) => Observable<InsightContent<unknown>>
 
     /**
-     * Returns content for the search based insight live preview chart.
-     */
-    getSearchInsightContent: (input: GetSearchInsightContentInput) => Promise<SeriesChartContent<unknown>>
-
-    /**
      * Returns content for the code stats insight live preview chart.
      */
     getLangStatsInsightContent: (input: GetLangStatsInsightContentInput) => Promise<CategoricalChartContent<unknown>>
@@ -111,9 +94,4 @@ export interface CodeInsightsBackend {
      * or "select:repo count:1" if no repository is found with the string "TODO"
      */
     getFirstExampleRepository: () => Observable<string>
-
-    /**
-     * Returns a features object used to show/hide and enable/disable UI elements
-     */
-    readonly UIFeatures: UiFeaturesConfig
 }