+++ client/web/src/repo/blob/Blob.tsx
@@ -185,7 +185,7 @@ const domFunctions = {
         if (!row) {
             throw new Error('Could not find closest row for codeCell')
         }
-        const numberCell = row.querySelector<HTMLTableCellElement>('td.line')
+        const numberCell = row.querySelector<HTMLTableCellElement>('th.line')
         if (!numberCell || !numberCell.dataset.line) {
             throw new Error('Could not find line number')
         }