+++ client/web/src/repo/blob/codemirror/tokens-as-links.ts
@@ -40,8 +40,18 @@ const focusSelectedLine = ViewPlugin.fromClass(
 
                 if (line) {
                     window.requestAnimationFrame(() => {
-                        const element = this.view.domAtPos(line.from).node as HTMLElement
-                        element.focus()
+                        const closestNode = this.view.domAtPos(line.from).node
+
+                        // Loosely find closest element.
+                        // Note: This is usually only be `closestNode` if the line is empty.
+                        const closestElement =
+                            closestNode instanceof HTMLElement ? closestNode : closestNode.parentElement
+
+                        const target = closestElement?.hasAttribute('data-line-focusable')
+                            ? closestElement
+                            : closestElement?.closest<HTMLElement>('[data-line-focusable]')
+
+                        target?.focus()
                     })
                 }
             }