+++ internal/codeintel/ranking/service_test.go
@@ -73,8 +73,9 @@ func TestGetDocumentRanks(t *testing.T) {
 	svc := newService(mockStore, nil, gitserverClient, nil, siteConfigQuerier{}, nil, &observation.TestContext)
 
 	mockStore.GetDocumentRanksFunc.SetDefaultReturn(map[string][2]float64{
-		"rust/main.rs": {0.00, 0.84},
-		"rust/lib.rs":  {0.25, 0.42},
+		"rust/main.rs": {1.00, 0.84},
+		"rust/lib.rs":  {0.75, 0.42},
+		"rust/min.js":  {0.25, 0.24}, // generated
 	}, true, nil)
 
 	gitserverClient.ListFilesForRepoFunc.SetDefaultReturn([]string{
@@ -100,23 +101,24 @@ func TestGetDocumentRanks(t *testing.T) {
 
 	expected := map[string][]float64{
 		// Precise
-		"rust/main.rs": {0.00, 0.84, 0, 0, 0, 0, 0},
-		"rust/lib.rs":  {0.25, 0.42, 0, 0, 0, 0, 0},
+		"rust/main.rs": {1.00, 1, 1, 1, 0.45652173, 1, 1}, // squashRange(0.84) -> 0.45652173
+		"rust/lib.rs":  {0.75, 1, 1, 1, 0.29577464, 1, 1}, // squashRange(0.42) -> 0.29577464
+		"rust/min.js":  {0.25, 0, 1, 1, 0.19354838, 1, 1}, // squashRange(0.24) -> 0.19354838
 
 		// Fallback
-		"code/a.go":           {1, 1, 1, 1, 1, 0.100, 1 - (0.00 / 13.0)},
-		"code/b.go":           {1, 1, 1, 1, 1, 0.100, 1 - (1.00 / 13.0)},
-		"code/c.go":           {1, 1, 1, 1, 1, 0.100, 1 - (2.00 / 13.0)},
-		"code/d.go":           {1, 1, 1, 1, 1, 0.100, 1 - (3.00 / 13.0)},
-		"main.go":             {1, 1, 1, 1, 1, 0.125, 1 - (4.00 / 13.0)},
-		"node_modules/bar.js": {1, 1, 1, 0, 1, 0.050, 1 - (5.00 / 13.0)},
-		"node_modules/baz.js": {1, 1, 1, 0, 1, 0.050, 1 - (6.00 / 13.0)},
-		"node_modules/foo.js": {1, 1, 1, 0, 1, 0.050, 1 - (7.00 / 13.0)},
-		"rendered/web/min.js": {1, 1, 0, 1, 1, 0.050, 1 - (8.00 / 13.0)},
-		"test/a.go":           {1, 1, 1, 1, 0, 0.100, 1 - (9.00 / 13.0)},
-		"test/b.go":           {1, 1, 1, 1, 0, 0.100, 1 - (10.0 / 13.0)},
-		"test/c.go":           {1, 1, 1, 1, 0, 0.100, 1 - (11.0 / 13.0)},
-		"test/d.go":           {1, 1, 1, 1, 0, 0.100, 1 - (12.0 / 13.0)},
+		"code/a.go":           {0, 1, 1, 1, 0, 0.100, 1 - (0.00 / 13.0)},
+		"code/b.go":           {0, 1, 1, 1, 0, 0.100, 1 - (1.00 / 13.0)},
+		"code/c.go":           {0, 1, 1, 1, 0, 0.100, 1 - (2.00 / 13.0)},
+		"code/d.go":           {0, 1, 1, 1, 0, 0.100, 1 - (3.00 / 13.0)},
+		"main.go":             {0, 1, 1, 1, 0, 0.125, 1 - (4.00 / 13.0)},
+		"node_modules/bar.js": {0, 1, 0, 1, 0, 0.050, 1 - (5.00 / 13.0)},
+		"node_modules/baz.js": {0, 1, 0, 1, 0, 0.050, 1 - (6.00 / 13.0)},
+		"node_modules/foo.js": {0, 1, 0, 1, 0, 0.050, 1 - (7.00 / 13.0)},
+		"rendered/web/min.js": {0, 0, 1, 1, 0, 0.050, 1 - (8.00 / 13.0)},
+		"test/a.go":           {0, 1, 1, 0, 0, 0.100, 1 - (9.00 / 13.0)},
+		"test/b.go":           {0, 1, 1, 0, 0, 0.100, 1 - (10.0 / 13.0)},
+		"test/c.go":           {0, 1, 1, 0, 0, 0.100, 1 - (11.0 / 13.0)},
+		"test/d.go":           {0, 1, 1, 0, 0, 0.100, 1 - (12.0 / 13.0)},
 	}
 
 	opt := cmp.Comparer(cmpFloat)